+++ platform/core/src/utils/metadataProvider/isEnhancedSOP.js
@@ -1,7 +1,10 @@
 import { sopClassDictionary } from '../sopClassDictionary';
 
 // ToDo: add support for other Enhanced SOPs
-const supportedEnhancedSOPs = [sopClassDictionary.EnhancedMRImageStorage];
+const supportedEnhancedSOPs = [
+  sopClassDictionary.EnhancedMRImageStorage,
+  sopClassDictionary.NuclearMedicineImageStorage,
+];
 
 const isEnhancedSOP = SOPClassUID => {
   return supportedEnhancedSOPs.includes(SOPClassUID);